+++ src/Exception/MiddlewareNotFound.php
@@ -1,6 +1,0 @@
-<?php
-namespace WoohooLabs\Harmony\Exception;
-
-class MiddlewareNotFound extends \Exception
-{
-}